Replaced Ifs With Function
Mapped If values and iterated through to determine tax bracket. Used @Aya example after understanding exactly what was happening in the For In If.

Also initialized x for the last For loop. Lint picked up that it may not have been defined. Didn't have an issue running without initializing x but I did it just to be safe.