Decided on CamelCase, Improved .format
Wasn't sure whether to go with snake or camel. Sticking with camel for
now and will try to keep var names short enough to still be descriptive.

Implemented formatting for numbers in order to display numbers as
accurately as they need to be (ie 12.3%, or $1,234.56) along with commas
to make larger numbers easier to interpret.